requirements.txt
streamlit==1.20.0
torch==1.13.1
torchvision==0.14.1
Pillow==9.4.0
transformers==4.26.1
scikit-learn==1.2.2
sentence-transformers==2.2.2
numpy==1.24.2
pandas==1.5.3
matplotlib==3.7.1
